fix(routes): Preserve /health/chat endpoint by moving to unprefixed router
- Created guardian/routes/health.py with unprefixed APIRouter
- Moved health_chat() handler from chat.py to health.py to prevent path conflicts
- Added base /health endpoint for system-level health checks
- Removed duplicate health_chat() from guardian_api.py
- Updated guardian_api.py to include health router without prefix

This ensures /health/chat remains accessible at the correct path for readiness
probes and load balancer health checks, preventing the route from becoming
/chat/health/chat when chat routes are mounted under /chat prefix.
