+++ guardian/routes/health.py
@@ -1,0 +1,43 @@
+"""
+Health Routes
+~~~~~~~~~~~~~
+
+Health check endpoints for monitoring subsystem status.
+Mounted without a prefix to preserve public paths like /health/chat.
+"""
+
+import logging
+from fastapi import APIRouter
+
+logger = logging.getLogger(__name__)
+
+# Import shared context
+try:
+    from guardian.guardian_api import chatlog_db, DB_BACKEND
+except ImportError as e:
+    logger.warning(f"[health] Import warning: {e}")
+    chatlog_db = None
+    DB_BACKEND = "unknown"
+
+
+# Create unprefixed router to preserve /health/chat path
+router = APIRouter(tags=["Health"])
+
+
+@router.get("/health")
+def health():
+    """Base health check endpoint for system-level monitoring."""
+    return {"status": "ok"}
+
+
+@router.get("/health/chat")
+def health_chat():
+    """Get health status of chat subsystem."""
+    try:
+        threads = chatlog_db.count_chat_threads()
+        messages = chatlog_db.count_all_messages()
+    except Exception as _e:
+        logger.warning("[health/chat] check failed: %s", _e)
+        threads = 0
+        messages = 0
+    return {"ok": True, "threads": threads, "messages": messages, "backend": DB_BACKEND}
